Make recursive function that pairs programmers and companies
The function cycles through the list of programmers and trys to match them with each of their preferences starting from the top. If a preference is taken it makes sure that they arn't more preferred than the currently paired person, if they are then it rematches the old pair. We also implemented a function to check for satisfactory and made minor adjustments to which company preferes which programmer.